+++ src/overlay/PeerRecord.h
@@ -25,7 +25,7 @@ class PeerRecord
 
     PeerRecord(){};
 
-    PeerRecord(string const& ip, uint32_t port,
+    PeerRecord(string const& ip, unsigned short port,
                VirtualClock::time_point nextAttempt, uint32_t fails,
                uint32_t rank)
         : mIP(ip)
@@ -44,14 +44,14 @@ class PeerRecord
                mNumFailures == other.mNumFailures && mRank == other.mRank;
     }
 
-    static void fromIPPort(std::string const& ip, uint32_t port,
+    static void fromIPPort(std::string const& ip, unsigned short port,
                            VirtualClock& clock, PeerRecord& ret);
     static bool parseIPPort(std::string const& ipPort, Application& app,
                             PeerRecord& ret,
-                            uint32_t defaultPort = DEFAULT_PEER_PORT);
+                            unsigned short defaultPort = DEFAULT_PEER_PORT);
 
     static optional<PeerRecord> loadPeerRecord(Database& db, std::string ip,
-                                               uint32_t port);
+                                               unsigned short port);
     static void loadPeerRecords(Database& db, uint32_t max,
                                 VirtualClock::time_point nextAttemptCutoff,
                                 vector<PeerRecord>& retList);
